backend/requirements.txt
Flask
Flask-GraphQL
graphene
tensorflow
scikit-learn
pandas
python-dotenv